rename: types 경로 명명 feature/TODAK-172/auth-integration
