+++ components/lottie-animation/MusicAnimation.tsx
@@ -2,7 +2,7 @@
 
 import Lottie from "lottie-react";
 import MusicAnimationData from "@/public/lottie-animation/music.json";
-import { AnimationType } from "@/utils/types";
+import { AnimationType } from "@/utils/types/common";
 
 const MusicAnimation: React.FC<AnimationType> = ({ style }) => {
   return (
