+++ utils/types/common/animationType.ts
